% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help.R
\name{get_help}
\alias{get_help}
\title{get help}
\usage{
get_help(pkg, fn = NULL)
}
\arguments{
\item{pkg}{where function to get help.}

\item{fn}{function name to get help. Default is NULL and raise errer.}
}
\description{
build help object from package
}
