% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help.R
\name{getfrom}
\alias{getfrom}
\alias{get_title}
\alias{get_description}
\alias{get_usage}
\title{get form help}
\usage{
get_title(help)

get_description(help)

get_usage(help)
}
\arguments{
\item{help}{help object from package \link{get_help}}
}
\description{
get form help
}
