% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usehelp-package.R
\docType{package}
\name{usehelp-package}
\alias{usehelp}
\alias{usehelp-package}
\title{usehelp: Use Help Text in Post, Presentation, etc}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
